refactor: extract WalletsOverlay from Dashboard
